remove unused code in appDelegate | sceneDelegate | delete main | viewController | organize folder
