+++ DailyCost/Application/SceneDelegate.swift
@@ -1,0 +1,33 @@
+//
+//  SceneDelegate.swift
+//  DailyCost
+//
+//  Created by Panji Yoga on 30/08/23.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let window = UIWindow(windowScene: windowScene)
+        let viewController = OnboardingViewController(nibName: "OnboardingViewController", bundle: nil)
+        window.rootViewController = viewController
+        self.window = window
+        window.makeKeyAndVisible()
+    }
+    
+    func sceneDidDisconnect(_ scene: UIScene) { }
+    
+    func sceneDidBecomeActive(_ scene: UIScene) { }
+    
+    func sceneWillResignActive(_ scene: UIScene) { }
+    
+    func sceneWillEnterForeground(_ scene: UIScene) { }
+    
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+    }
+}
