+++ DailyCost/SupportingFiles/Assets.xcassets/AccentColor.colorset/Contents.json
