+++ DailyCost/SupportingFiles/Assets.xcassets/Contents.json
